requirements.txt
streamlit==1.32.2
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
Pillow==10.2.0
easyocr
torch
torchvision
opencv-python-headless
